src/VFR.AiEngine/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
celery==5.4.0
redis==5.0.4
python-multipart==0.0.9
torch==2.3.0
smplx
trimesh==4.4.0
scipy==1.13.0
numpy==1.26.4
